+++ spec/spec_helper.rb
@@ -8,6 +8,7 @@ require 'parallel_tests'
 require 'parallel_tests/test/runtime_logger'
 require 'parallel_tests/rspec/runtime_logger'
 require 'parallel_tests/rspec/summary_logger'
+require 'parallel_tests/rspec/verbose_formatter'
 
 String.class_eval do
   def strip_heredoc